Move lightclient circuit input and output to rollup interface
