+++ crates/light-client-prover/src/lib.rs
@@ -1,7 +1,4 @@
 pub mod circuit;
-pub mod input;
-pub mod output;
-
 #[cfg(feature = "native")]
 pub mod da_block_handler;
 #[cfg(feature = "native")]
