--- crates/light-client-prover/src/output.rs
@@ -1,6 +1,0 @@
-use borsh::{BorshDeserialize, BorshSerialize};
-
-#[derive(Debug, Clone, BorshDeserialize, BorshSerialize)]
-pub struct LightClientCircuitOutput {
-    pub state_root: [u8; 32],
-}
